Show the importance of the git history
We really value both the commit history as well as well written commit
messages. Messages can be as long as the committer wants, but the first
line can not exceed 50 characters. The next lines should not exceed 72
characters. This is the optimal width for any display and github.

A good commit message should not just describe _what_ happened, but also
_why_ it happened. Specially after a couple of months, that is the most
important information a commit message can provide, wouldn't you agree?